Спринт 14 на провеку поправил ProgressBar и исправил classPlayer
